add error handler to some routes complete the swagger

+++ src/routes/pokemonRouter.js
@@ -2,6 +2,7 @@ import express from 'express';
 import Pokemon from '../model/Pokemon.js';
 import checkAuthorization from '../middleware/checkAuthorization.js';
 import isAdmin from '../middleware/isAdmin.js';
+import Trainer from "../model/Trainer.js";
 
 const pokemonRouter = express.Router();
 
@@ -17,6 +18,9 @@ pokemonRouter.post('/add/self', [checkAuthorization, async (req, res) => {
   } = req.body;
 
   try {
+    const trainer = await Trainer.findByPk(res.locals.requestor.id)
+    if (!trainer) return res.status(404).send('Trainer not found')
+
     const { id } = await Pokemon.create({
       species,
       name,
@@ -48,6 +52,9 @@ pokemonRouter.post('/add/:trainerId', [checkAuthorization, isAdmin, async (req, 
   const { trainerId } = req.params;
 
   try {
+    const trainer = await Trainer.findByPk(res.locals.requestor.id)
+    if (!trainer) return res.status(404).send('Trainer not found')
+
     const { id } = await Pokemon.create({
       species,
       name,
@@ -67,8 +74,8 @@ pokemonRouter.post('/add/:trainerId', [checkAuthorization, isAdmin, async (req, 
 
 pokemonRouter.get('/getPokemons/self', [checkAuthorization, async (req, res) => {
   try {
-    const { id } = await Pokemon.findAll({ where: { trainerId: res.locals.requestor.id } });
-    return res.status(200).send({ id });
+    const pokemon = await Pokemon.findAll({ where: { trainerId: res.locals.requestor.id } });
+    return res.status(200).send(pokemon);
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -77,8 +84,8 @@ pokemonRouter.get('/getPokemons/self', [checkAuthorization, async (req, res) => 
 pokemonRouter.get('/getPokemons/:trainerId', [checkAuthorization, async (req, res) => {
   const { trainerId } = req.params;
   try {
-    const { id } = await Pokemon.findAll({ where: { trainerId } });
-    return res.status(200).send({ id });
+    const pokemon = await Pokemon.findAll({ where: { trainerId } });
+    return res.status(200).send(pokemon);
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -88,10 +95,11 @@ pokemonRouter.delete('/release/self', [checkAuthorization, async (req, res) => {
   const { id } = req.body;
   try {
     const pokemon = await Pokemon.findByPk(id);
+    if (!pokemon) return res.status(404).send('pokemon not found')
     if (pokemon.trainerId === res.locals.requestor.id) {
       await Pokemon.destroy({ where: { id } });
-      return res.status(200).send('account deleted');
-    } return res.status(401).send('this is not your pokemon!');
+      return res.status(200).send('pokemon released');
+    } return res.status(400).send('this is not your pokemon!');
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -102,10 +110,11 @@ pokemonRouter.delete('/release/:trainerId', [checkAuthorization, isAdmin, async 
   const { id } = req.body;
   try {
     const pokemon = await Pokemon.findByPk(id);
+    if (!pokemon) return res.status(404).send('pokemon not found')
     if (pokemon.trainerId === trainerId) {
       await Pokemon.destroy({ where: { id } });
       return res.status(200).send('account deleted');
-    } return res.status(401).send('the trainer don\'t own this pokemon');
+    } return res.status(400).send('the trainer don\'t own this pokemon');
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -123,7 +132,7 @@ pokemonRouter.patch('/rename/self', [checkAuthorization, async (req, res) => {
       const updatedPokemon = await pokemon.update(pokemonToUpdate);
 
       return res.status(200).send(updatedPokemon);
-    } return res.status(401).send('this is not your pokemon!');
+    } return res.status(400).send('this is not your pokemon!');
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -143,7 +152,7 @@ pokemonRouter.patch('/rename/:trainerId', [checkAuthorization, isAdmin, async (r
       const updatedPokemon = await pokemon.update(pokemonToUpdate);
 
       return res.status(200).send(updatedPokemon);
-    } return res.status(401).send('the trainer don\'t own this pokemon');
+    } return res.status(400).send('the trainer don\'t own this pokemon');
   } catch (error) {
     return res.status(500).send(error);
   }
@@ -163,7 +172,7 @@ pokemonRouter.patch('/levelUp/:trainerId', [checkAuthorization, isAdmin, async (
       const updatedPokemon = await pokemon.update(pokemonToUpdate);
 
       return res.status(200).send(updatedPokemon);
-    } return res.status(401).send('the trainer don\'t own this pokemon');
+    } return res.status(400).send('the trainer don\'t own this pokemon');
   } catch (error) {
     return res.status(500).send(error);
   }
